requirements.txt
pandas==2.1.0
numpy==1.24.3
yfinance==0.2.28
scikit-learn==1.3.0
ta==0.10.2
praw==7.7.1
newsapi-python==0.2.7
textblob==0.17.1
pandas-ta==0.3.14b
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.27.0
python-dotenv==1.0.0
tensorflow==2.13.0
plotly==5.16.1
requests==2.31.0 